requirements.txt
ipykernel==5.2.0
jupyter==1.0.0
matplotlib==3.2.1
nbconvert==5.6.1
numpy==1.18.2
opencv-python==4.2.0.32
pandas==1.0.3
Pillow==7.0.0
plotly==4.9.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
tqdm==4.46.0
trimesh==3.6.16
imageio==2.9.0
ninja==1.10.0
imgaug==0.4.0
lpips==0.1.3
moviepy==1.0.3
munch==2.5.0
yamlenv==0.7.1
omegaconf==2.0.6
smplx==0.1.26
efficientnet-pytorch==0.7.0
kornia

git+https://github.com/karfly/nvdiffrast_compute-capability_6.0
git+https://github.com/nghorbani/configer
